Updating Apex to OpenStack Ocata
- power management updated to virtualbmc, pxe_ssh is deprecated
- removing custom tacker build
- removing custom congress build
- disabling yum update in undercloud on the cli instead of in a patch
- Undercloud is direct kernel booted now, there are no kernel and initrd
  in the disk image from upstream
- remove OpenDaylight previous to Carbon

JIRA: APEX-433
JIRA: APEX-432
JIRA: APEX-431

+++ lib/overcloud-deploy-functions.sh
@@ -40,14 +40,14 @@ function overcloud_deploy {
         DEPLOY_OPTIONS+=" -e /usr/share/openstack-tripleo-heat-templates/environments/neutron-opendaylight-honeycomb-l2.yaml"
       fi
     else
-      DEPLOY_OPTIONS+=" -e /usr/share/openstack-tripleo-heat-templates/environments/neutron-opendaylight-l3.yaml"
+      DEPLOY_OPTIONS+=" -e /usr/share/openstack-tripleo-heat-templates/environments/neutron-opendaylight.yaml"
     fi
     SDN_IMAGE=opendaylight
   elif [ "${deploy_options_array['sdn_controller']}" == 'opendaylight-external' ]; then
     DEPLOY_OPTIONS+=" -e /usr/share/openstack-tripleo-heat-templates/environments/opendaylight-external.yaml"
     SDN_IMAGE=opendaylight
   elif [ "${deploy_options_array['sdn_controller']}" == 'onos' ]; then
-    echo -e "${red}ERROR: ONOS is unsupported in Danube...exiting${reset}"
+    echo -e "${red}ERROR: ONOS is unsupported in Euphrates...exiting${reset}"
     exit 1
     #if [ "${deploy_options_array['sfc']}" == 'True' ]; then
     #  DEPLOY_OPTIONS+=" -e /usr/share/openstack-tripleo-heat-templates/environments/onos_sfc.yaml"
@@ -266,22 +266,20 @@ EOI
   # Set ODL version accordingly
   if [[ "${deploy_options_array['sdn_controller']}" == 'opendaylight' && -n "${deploy_options_array['odl_version']}" ]]; then
     case "${deploy_options_array['odl_version']}" in
-      beryllium) odl_version=''
+      carbon) odl_version=''
               ;;
-      boron)  odl_version='boron'
-              ;;
-      carbon) odl_version='master'
-              ;;
-      *) echo -e "${red}Invalid ODL version ${deploy_options_array['odl_version']}.  Please use 'carbon' or 'boron' values.${reset}"
+      *) echo -e "${red}Invalid ODL version ${deploy_options_array['odl_version']}.  Please use 'carbon'.${reset}"
          exit 1
          ;;
     esac
 
-    ssh -T ${SSH_OPTIONS[@]} "stack@$UNDERCLOUD" <<EOI
-      LIBGUESTFS_BACKEND=direct virt-customize --run-command "yum -y remove opendaylight" \
-                                               --run-command "yum -y install /root/${odl_version}/*" \
-                                               -a overcloud-full.qcow2
+    if [[ "$odl_version" ]]; then
+        ssh -T ${SSH_OPTIONS[@]} "stack@$UNDERCLOUD" <<EOI
+          LIBGUESTFS_BACKEND=direct virt-customize --run-command "yum -y remove opendaylight" \
+                                                   --run-command "yum -y install /root/${odl_version}/*" \
+                                                   -a overcloud-full.qcow2
 EOI
+    fi
   fi
 
   # Override ODL if FDIO and ODL L2
@@ -352,6 +350,9 @@ EOI
   DEPLOY_OPTIONS+=" --control-flavor control --compute-flavor compute"
   if [[ "$virtual" == "TRUE" ]]; then
      DEPLOY_OPTIONS+=" -e virtual-environment.yaml"
+     # double check the status of the vbmc devices
+     # TODO add some validation logic here
+     vbmc list
   fi
 
   echo -e "${blue}INFO: Deploy options set:\n${DEPLOY_OPTIONS}${reset}"
@@ -375,33 +376,25 @@ echo "Uploading overcloud glance images"
 openstack overcloud image upload
 
 echo "Configuring undercloud and discovering nodes"
-openstack baremetal import --json instackenv.json
+
 
 if [[ -z "$virtual" ]]; then
-  openstack baremetal introspection bulk start
-  if [[ -n "$root_disk_list" ]]; then
-    openstack baremetal configure boot --root-device=${root_disk_list}
-  else
-    openstack baremetal configure boot
-  fi
+  openstack overcloud node import instackenv.json
+  openstack overcloud node introspect --all-manageable --provide
+  #if [[ -n "$root_disk_list" ]]; then
+    # TODO: replace node configure boot with ironic node-update
+    # TODO: configure boot is not used in ocata here anymore
+    #openstack overcloud node configure boot --root-device=${root_disk_list}
+    #https://github.com/openstack/tripleo-quickstart-extras/blob/master/roles/overcloud-prep-images/templates/overcloud-prep-images.sh.j2#L73-L130
+    #ironic node-update $ironic_node add properties/root_device='{"{{ node['key'] }}": "{{ node['value'] }}"}'
+  #fi
 else
-  openstack baremetal configure boot
+  openstack overcloud node import --provide instackenv.json
 fi
 
-echo "Configuring flavors"
-for flavor in baremetal control compute; do
-  echo -e "${blue}INFO: Updating flavor: \${flavor}${reset}"
-  if openstack flavor list | grep \${flavor}; then
-    openstack flavor delete \${flavor}
-  fi
-  openstack flavor create --id auto --ram 4096 --disk 39 --vcpus 1 \${flavor}
-  if ! openstack flavor list | grep \${flavor}; then
-    echo -e "${red}ERROR: Unable to create flavor \${flavor}${reset}"
-  fi
-done
-openstack flavor set --property "cpu_arch"="x86_64" --property "capabilities:boot_option"="local" baremetal
-openstack flavor set --property "cpu_arch"="x86_64" --property "capabilities:boot_option"="local" --property "capabilities:profile"="control" control
-openstack flavor set --property "cpu_arch"="x86_64" --property "capabilities:boot_option"="local" --property "capabilities:profile"="compute" compute
+openstack flavor set --property "cpu_arch"="x86_64" baremetal
+openstack flavor set --property "cpu_arch"="x86_64" control
+openstack flavor set --property "cpu_arch"="x86_64" compute
 echo "Configuring nameserver on ctlplane network"
 dns_server_ext=''
 for dns_server in ${dns_servers}; do
@@ -409,7 +402,7 @@ for dns_server in ${dns_servers}; do
 done
 neutron subnet-update \$(neutron subnet-list | grep -Ev "id|tenant|external|storage" | grep -v \\\\-\\\\- | awk {'print \$2'}) \${dns_server_ext}
 sed -i '/CloudDomain:/c\  CloudDomain: '${domain_name} ${ENV_FILE}
-echo "Executing overcloud deployment, this should run for an extended period without output."
+echo "Executing overcloud deployment, this could run for an extended period without output."
 sleep 60 #wait for Hypervisor stats to check-in to nova
 # save deploy command so it can be used for debugging
 cat > deploy_command << EOF
